Allow reading unaligned pointers in ObjectTypeInformation fields
The ObjectTypeInformation `type_name` field may be returned as an
unaligned pointer. This will cause a debug assertion to fail when
built with Rust 1.71 or higher: Rust 1.71 added debug assertions
that panic when reading unaligned pointers since unaligned pointers
may lead to undefined behavior.

We assume that the pointers given by the NtQueryObject call are
valid despite being unaligned since the function returns the
STATUS_SUCCESS code. This is similar to places in the Rust standard
library that handle cases of Windows returning unexpectedly unaligned
pointers. For example:
https://github.com/rust-lang/rust/blob/f91b634643de14ab5156b0c084d80bd6845fb0ae/library/std/src/sys/windows/fs.rs#L765-L782